consolidate ReleaseTransactionBuilder tests and assertion helpers - Keep a single ReleaseTransactionBuilderTest with @Nested suites per API   (buildAmountTo, buildEmptyWalletTo, buildMigrationTransaction, batched   peg-outs, mock/regtest paths). - Inline former ReleaseTransactionBuilderAssertions at the bottom of the   test class; remove the standalone assertions file. - Add composite input checks (e.g. assertReleaseTxInputs*,   assertMigrationReleaseTxInputs*) and ticket-style output helpers   (assertOutputsWithNoChange / assertOutputsWithNonDustChange /   assertOutputsWithDustChange), using qualified calls where instance   overloads shadow static helpers. - Delete superseded standalone test classes after merging their cases.
